ALSA: timer: Fix negative queue usage by racy accesses
commit 3fa6993fef634e05d200d141a85df0b044572364 upstream.

The user timer tu->qused counter may go to a negative value when
multiple concurrent reads are performed since both the check and the
decrement of tu->qused are done in two individual locked contexts.
This results in bogus read outs, and the endless loop in the
user-space side.

The fix is to move the decrement of the tu->qused counter into the
same spinlock context as the zero-check of the counter.

Note: fx30-3g, 0019-ALPC-374-CVE-2017-1000380.patch

Resolves: LXSWIREF-1354
Change-Id: If9f97c869e2490021b2e3e0e3315c512d9ed43ca
Signed-off-by: Takashi Iwai <tiwai@suse.de>
Signed-off-by: Ben Hutchings <ben@decadent.org.uk>
Signed-off-by: Dragan Marinkovic <dmarinkovi@sierrawireless.com>