[MangOH] Include drivers for IoT expansion slots
AR7 kernel does not assume existence of IoT slots that are present
on the similar MangOH platform. In order to map the IoT slots, the
kernel needs to support I2C muxing, PCA9548 I2C switch, SX1509
GPIO expanders and I2C EEPROMs on IoT cards. Presence of the IoT
slots (and MangOH platform) still need to be determined by
inserting a MangOH-specific kernel module, which will be provided
with Legato. This commit only adds necessary drivers to perform
IoT slot mapping.

Note that the current kernel version supports SX150X devices as
built-in and hence the decision was brought not to add the above
drivers as modules in order to maintain consistency. Assumption is
that the above drivers are simple enough not to substantionally
increase kernel size.

Also build spidev as a module to avoid driver binding at boot time.

Resolves: LXSWI9X1517-51
Change-Id: I29f11cb1644bb961101b557f82d1adf09006108a
Signed-off-by: Zoran Markovic <zmarkovic@sierrawireless.com>