ALSA: timer: Reject user params with too small ticks
commit 71321eb3f2d0df4e6c327e0b936eec4458a12054 upstream.

When a user sets a too small ticks with a fine-grained timer like
hrtimer, the kernel tries to fire up the timer irq too frequently.
This may lead to the condensed locks, eventually the kernel spinlock
lockup with warnings.

For avoiding such a situation, we define a lower limit of the
resolution, namely 1ms.  When the user passes a too small tick value
that results in less than that, the kernel returns -EINVAL now.

Note: fx30-3g, 0019-ALPC-374-CVE-2017-1000380.patch

Resolves: LXSWIREF-1354
Change-Id: I11f06ecd9091413a006cb19859adad5d4a72ad13
Reported-by: Dmitry Vyukov <dvyukov@google.com>
Signed-off-by: Takashi Iwai <tiwai@suse.de>
Signed-off-by: Ben Hutchings <ben@decadent.org.uk>
Signed-off-by: Dragan Marinkovic <dmarinkovi@sierrawireless.com>