swimcu: improvements in shutdown procedure
Currently, the shutdown procedure for ULPM does not terminate tasks
correctly, and imposes a risk of file corruption. This change adds
support for new MCU commands to set a sync time after ULPM has been
enabled. This sync period allows the system to terminate all
applications and processes, and unmount file systems before poweroff.

When a ULPM request is received, the swimcu driver validates the wakeup
sources and sends the sync time message to the MCU before invoking
/sbin/poweroff. Once the shutdown is complete, swimcu sends the final
sync message to the MCU to notify that it is safe to remove power. If
the MCU does not receive this message within the sync time period, it
will remove power anyway.

If /sbin/poweroff is invoked directly by the user, swimcu will configure
ULPM after shutdown is complete and set the sync timer to zero to notify
the MCU that power can be removed immediately.

Resolves: LXSWI9X1517-247
Change-Id: Ic6ad3865ee3f0b6595b4e2bb13d8257124a41bb0
Signed-off-by: Mitchell Nichols <MNichols@sierrawireless.com>