wl18xx: reset ECPU across suspend/resume cycle
According to:
http://e2e.ti.com/support/wireless-connectivity/wifi/f/968/t/181835
the WLAN portion of the chip has to be reset each time the firmware
needs restart. This is particularly a problem across suspend/resume
cycle after which the firmware has to be re-loaded due to internal
RAM being powered down.

Force a reset of platform-provided WLAN_EN in the above case to
facilitate firmware restart.

Resolves: LXSWIREF-1056
Change-Id: I8d7daa67cf2664a30b52a162cfee46a7715364a0
Signed-off-by: Zoran Markovic <zmarkovic@sierrawireless.com>