gpio: export and unexport does not work properly
This fixes the bug where the following pairs of export/unexport
functions are not working properly:

    /sys/class/gpio/export
    /sys/class/gpio/unexport

    /sys/class/gpio/v2/export
    /sys/class/gpio/v2/unexport

    /sys/class/gpio/v2/alias_export
    /sys/class/gpio/v2/alias_unexport

Entry pairs should not be mixed. For example, it is not
recommended using /sys/class/gpio/export to export GPIO
pin and than using /sys/class/gpio/v2/unexport to
remove that pin from the exported list.

Change-Id: I1359db4dcb75b3ad3206b871cf1dc9c52dd6e5ac
Resolves: LXSWIREF-1632
Signed-off-by: Dragan Marinkovic <dmarinkovi@sierrawireless.com>